Input handling
Unsure of the best way to go about having for
instance 9 + 9 - 10, I understand I take 9+9 and then
subtract total from 10 but do I make a curr var hold
the answer or set the answer equal to var1 for the
next operation. I think the ladder so I've implemented
a test of it on multiply.