Return value added
So I'm slightly confused on the scope of my variables
which are being changed within a function within
a function. So I've decided to just haave my arthimetic
functions return a value which is returned to val1.
It works properly before it just returned val1 as if
no operation had occured. This prob isnt a permanent
fix but works for now!